Reworked previous scripts for modularity and integrated into Week 6 Turn to Menu file.

+++ Week2/Process Management #4.ps1
@@ -1,8 +1,20 @@
-﻿if (Get-Process -Name "chrome" -ErrorAction SilentlyContinue) {
-    Stop-Process -Name "chrome" -Force
-    Write-Host "Google Chrome was running and has been stopped."
+﻿function openAndClose() {
+    if (Get-Process -Name "chrome" -ErrorAction SilentlyContinue) {
+            Stop-Process -Name "chrome" -Force
+            Write-Host "Google Chrome was running and has been stopped."
+    }
+    else {
+            Start-Process "chrome.exe" "https://www.champlain.edu/"
+            Write-Host "Google Chrome was not running. Starting now and directing to champlain.edu"
+    } 
 }
-else {
-    Start-Process "chrome.exe" "https://www.champlain.edu/"
-    Write-Host "Google Chrome was not running. Starting now and directing to champlain.edu"
+
+function champlainEduOnlyOpen() {    
+    $checkChrome = Get-Process -Name "chrome" -ErrorAction SilentlyContinue 
+    if (!$checkChrome) {
+        Start-Process "chrome.exe" "https://www.champlain.edu/"
+        Write-Host "Google Chrome was not running. Starting now and directing to champlain.edu" 
+    } else {
+        Write-Host "Chrome is already running. No action taken." 
+    }
 }
